[Add]부서별 직원 수 로직 생성 중
