+++ bhr/src/main/java/com/nineties/bhr/admin/dto/AdminDTO.java
@@ -1,5 +1,6 @@
-package com.nineties.bhr.emp.dto;
+package com.nineties.bhr.admin.dto;
 
+import com.nineties.bhr.dept.domain.Dept;
 import lombok.Getter;
 import lombok.Setter;
 
@@ -9,4 +10,5 @@ public class AdminDTO {
     //관리자페이지에서 불러올 DTO생성
     private Long empNo;
     private String name;
+    private Dept dept;
 }
