+++ bhr/src/main/java/com/nineties/bhr/admin/dto/deptDTO.java
@@ -1,0 +1,13 @@
+package com.nineties.bhr.admin.dto;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class deptDTO {
+    private Long empNo;
+    private Long id;
+    private String deptName;
+    private int numOfEmployeesInDept;
+}
