+++ bhr/src/main/java/com/nineties/bhr/emp/repository/EmployeesRepository.java
@@ -4,7 +4,6 @@ import com.nineties.bhr.emp.domain.Employees;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.util.List;
 import java.util.Optional;
 
 @Repository
